Update MC_York.m
Changed output of MC_York to report the coefficient and exponent of the fit relationship (as opposed to the coefficient and 1 over the exponent of the fit relationship). Added logical flag to suppress drawing of figure if desired (and added error bars to points). Removed use of nanmean and provided option to fix b directly as opposed to it being based on the length of the b input.